requirements.txt
opencv_python
Requests
PyYAML
fake-useragent